pyproject.toml
[project]
name = "rag-with-azure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.25.2",
    "azure-search-documents>=11.6.0",
    "azure-storage-blob>=12.28.0",
    "beautifulsoup4>=4.14.3",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
